split extension for participant-verifier
