+++ extensions/participant-verifier/src/main/java/org/eclipse/dataspaceconnector/registration/verifier/DefaultParticipantVerifier.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.registration.authority;
+package org.eclipse.dataspaceconnector.registration.verifier;
 
 import org.eclipse.dataspaceconnector.iam.did.spi.credentials.CredentialsVerifier;
 import org.eclipse.dataspaceconnector.iam.did.spi.resolution.DidResolverRegistry;
