+++ launcher/build.gradle.kts
@@ -25,6 +25,7 @@ val identityHubGroup: String by project
 
 dependencies {
     implementation(project(":extensions:registration-service"))
+    implementation(project(":extensions:participant-verifier"))
     implementation("${edcGroup}:identity-did-web:${edcVersion}")
     implementation("${edcGroup}:identity-did-core:${edcVersion}")
     implementation("${edcGroup}:core:${edcVersion}")
