+++ settings.gradle.kts
@@ -2,6 +2,7 @@ rootProject.name = "registration-service"
 
 include(":launcher")
 include(":extensions:registration-service")
+include(":extensions:participant-verifier")
 include(":extensions:participant-store-spi")
 include(":extensions:dataspace-authority-spi")
 include(":system-tests")
